clean up interceptor API & add more example uses
