+++ Samples/SampleApp/Sources/AppDelegate.swift
@@ -15,6 +15,7 @@
  */
 
 import UIKit
+import Workflow
 import WorkflowUI
 
 @UIApplicationMain
@@ -24,7 +25,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         window = UIWindow(frame: UIScreen.main.bounds)
 
-        window?.rootViewController = WorkflowHostingController(workflow: RootWorkflow())
+        let interceptors: [WorkflowInterceptor] = [
+            RootRenderPassTimer(),
+            SimpleActionLogger(),
+        ]
+
+        window?.rootViewController = WorkflowHostingController(workflow: RootWorkflow(), interceptors: interceptors)
 
         window?.makeKeyAndVisible()
