+++ Workflow/Sources/WorkflowNode.swift
@@ -160,255 +160,3 @@ extension WorkflowNode {
         var debugInfo: WorkflowUpdateDebugInfo
     }
 }
-
-// MARK: - Experimental Observer Stuff
-
-public protocol RenderContextInterceptor {
-    // would this be useful?
-}
-
-public protocol WorkflowInterceptor {
-    func onSessionStarted(
-        _ session: WorkflowSession
-    )
-
-    func onMakeInitialState<
-        W: Workflow
-    >(
-        workflow: W,
-        proceed: (W) -> W.State,
-        session: WorkflowSession
-    ) -> W.State
-
-    func onRender<
-        W: Workflow
-    >(
-        workflow: W,
-        state: W.State,
-        proceed: (W, W.State) -> W.Rendering,
-        session: WorkflowSession
-    ) -> W.Rendering
-
-    func onActionSent<Action: WorkflowAction>(
-        action: Action,
-        proceed: (Action) -> Void
-    )
-}
-
-// no-op interface
-public protocol NoOpWorkflowInterceptor: WorkflowInterceptor {}
-
-public extension NoOpWorkflowInterceptor {
-    func onSessionStarted(_ session: WorkflowSession) {}
-
-    func onMakeInitialState<
-        W: Workflow
-    >(
-        workflow: W,
-        proceed: (W) -> W.State,
-        session: WorkflowSession
-    ) -> W.State {
-        proceed(workflow)
-    }
-
-    func onRender<
-        W: Workflow
-    >(
-        workflow: W,
-        state: W.State,
-        proceed: (W, W.State) -> W.Rendering,
-        session: WorkflowSession
-    ) -> W.Rendering {
-        proceed(workflow, state)
-    }
-
-    func onActionSent<Action: WorkflowAction>(
-        action: Action,
-        proceed: (Action) -> Void
-    ) {
-        proceed(action)
-    }
-}
-
-struct NoOpWorkflowInterceptorImpl: NoOpWorkflowInterceptor {
-    init() {}
-}
-
-public class WorkflowInterceptorImpl: NoOpWorkflowInterceptor {
-    public func onRender<W: Workflow>(
-        workflow: W,
-        state: W.State,
-        proceed: (W, W.State) -> W.Rendering,
-        session: WorkflowSession
-    ) -> W.Rendering {
-        print("[interceptor]: about to render")
-        defer { print("[interceptor]: rendered") }
-
-        return proceed(workflow, state)
-    }
-}
-
-/**/
-public class WorkflowSession {
-    private static var _nextID: UInt64 = 0
-    static func makeSessionID() -> UInt64 {
-        _nextID += 1
-        return _nextID
-    }
-
-    public let typeDescriptor: String
-
-    public let renderKey: String
-
-    public let sessionID: UInt64
-
-    public let parent: WorkflowSession?
-
-    init<WorkflowType: Workflow>(
-        workflow: WorkflowType,
-        renderKey: String,
-        parent: WorkflowSession?
-    ) {
-        self.typeDescriptor = String(describing: workflow.self)
-        self.renderKey = renderKey
-        self.sessionID = Self.makeSessionID()
-        self.parent = parent
-    }
-}
-
-struct ChainedWorkflowInterceptor: WorkflowInterceptor {
-    private let interceptors: [WorkflowInterceptor]
-
-    init(interceptors: [WorkflowInterceptor]) {
-        self.interceptors = interceptors
-    }
-
-    func onSessionStarted(_ session: WorkflowSession) {
-        interceptors.forEach {
-            $0.onSessionStarted(session)
-        }
-    }
-
-    func onMakeInitialState<W>(
-        workflow: W,
-        proceed: (W) -> W.State,
-        session: WorkflowSession
-    ) -> W.State where W: Workflow {
-        let result: W.State = withoutActuallyEscaping(proceed) { proceed in
-            let chainedProceed = interceptors.reduce(proceed) { proceedAccum, interceptor in
-                { workflow in
-                    interceptor.onMakeInitialState(
-                        workflow: workflow,
-                        proceed: proceedAccum,
-                        session: session
-                    )
-                }
-            }
-
-            return chainedProceed(workflow)
-        }
-
-        return result
-    }
-
-    func onRender<W>(
-        workflow: W,
-        state: W.State,
-        proceed: (W, W.State) -> W.Rendering,
-        session: WorkflowSession
-    ) -> W.Rendering where W: Workflow {
-        let rez: W.Rendering = withoutActuallyEscaping(proceed) { proceed in
-            let chainedProceed = interceptors.reduce(proceed) { partialResult, interceptor in
-                { workflow, state in
-                    interceptor.onRender(
-                        workflow: workflow,
-                        state: state,
-                        proceed: partialResult,
-                        session: session
-                    )
-                }
-            }
-
-            let rendering = chainedProceed(workflow, state)
-            return rendering
-        }
-
-        return rez
-    }
-
-    func onActionSent<Action>(
-        action: Action,
-        proceed: (Action) -> Void
-    ) where Action: WorkflowAction {
-        withoutActuallyEscaping(proceed) { proceed in
-            let chainedProceed = interceptors.reduce(proceed) { proceedAccum, interceptor in
-                { action in
-                    interceptor.onActionSent(
-                        action: action,
-                        proceed: proceedAccum
-                    )
-                }
-            }
-
-            chainedProceed(action)
-        }
-    }
-}
-
-extension Array where Element == WorkflowInterceptor {
-    func chained() -> ChainedWorkflowInterceptor {
-        ChainedWorkflowInterceptor(interceptors: self)
-    }
-}
-
-protocol LoggableAction {
-    var loggingDescription: String { get }
-}
-
-struct RootRenderPassTimer: NoOpWorkflowInterceptor {
-    func onRender<W>(workflow: W, state: W.State, proceed: (W, W.State) -> W.Rendering, session: WorkflowSession) -> W.Rendering where W: Workflow {
-        if session.parent == nil {
-            let tock = CACurrentMediaTime()
-            defer {
-                let tick = CACurrentMediaTime()
-                let renderDuration = tick - tock
-
-                print("[interceptor]: root render duration: \(renderDuration)")
-            }
-
-            return proceed(workflow, state)
-        } else {
-            return proceed(workflow, state)
-        }
-    }
-}
-
-struct SimpleActionLogger: NoOpWorkflowInterceptor {
-    let log: (String) -> Void = {
-        print("[interceptor]: " + $0)
-    }
-
-    func onActionSent<Action>(
-        action: Action,
-        proceed: (Action) -> Void
-    ) where Action: WorkflowAction {
-        switch action {
-        case let action as LoggableAction:
-            // TODO: maybe there is a way to avoid dynamic casting
-            log("got loggable action: \(action.loggingDescription)")
-        default:
-            log("got default action of type: \(String(describing: action.self))")
-        }
-        proceed(action)
-    }
-}
-
-// counts number of nodes created over time (in a single tree)
-final class SimpleSessionCounter: NoOpWorkflowInterceptor {
-    var sessionCount = 0
-
-    func onSessionStarted(_ session: WorkflowSession) {
-        sessionCount += 1
-        print("session count: \(sessionCount)")
-    }
-}
